requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cssselect==1.3.0
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
idna==3.10
ipykernel==6.29.5
ipython==8.18.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba3k==0.35.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==4.9.3
lxml_html_clean==0.1.0
matplotlib-inline==0.1.7
multitasking==0.0.11
mysql-connector-python==9.2.0
names-dataset==3.3.1
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
numpy>=1.26.0,<2.0.0
packaging==24.2
pandas>=2.2.1
parso==0.8.4
pexpect==4.9.0
Pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycountry==24.6.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tinysegmenter==0.3
tldextract==5.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
yfinance==0.2.37
